requirements.txt
click>=8.0.0
rich>=12.0.0
pyyaml>=6.0
requests>=2.28.0
watchdog>=2.1.0
gitpython>=3.1.0
jsonschema>=4.0.0
jinja2>=3.1.0
colorama>=0.4.0
python-dateutil>=2.8.0
pathspec>=0.10.0